1. pifComm_AttachTask 함수에서 task는 항상 정지 상태로 설정한다. 이후 pifComm_AttachClient함수가 호출될 때 task를 실행한다. 또한 pifComm_DetachClient 함수가 호출되면 task는 정지한다.
2. pifLog_PrintChar 함수 추가.
3. pifLog 기능 개선.
4. pifGps에서 Error 발생시 evt_nmea_receive를 호출하도록 함.
5. Basic interpreter인 pifBasic 기능 추가.

+++ source/sensor/pif_hmc5883.c
@@ -130,7 +130,7 @@ BOOL pifHmc5883_Init(PifHmc5883* p_owner, PifId id, PifI2cPort* p_i2c, PifImuSen
     }
 
 #ifndef __PIF_NO_LOG__
-    pifLog_Printf(LT_INFO, "Mag scale: X=%f Y=%f Z=%f", p_owner->scale[AXIS_X], p_owner->scale[AXIS_Y], p_owner->scale[AXIS_Z]);
+    pifLog_Printf(LT_INFO, "Mag scale: X=%f Y=%f Z=%f", (double)p_owner->scale[AXIS_X], (double)p_owner->scale[AXIS_Y], (double)p_owner->scale[AXIS_Z]);
 #endif
 
 	p_imu_sensor->_measure |= IMU_MEASURE_MAGNETO;
